(intl/database) Apply exhentai database and traslate some text

+++ lib/pages/database_download/database_download_page.dart
@@ -39,8 +39,10 @@ class DataBaseDownloadPage extends StatefulWidget {
   final bool isExistsDataBase;
   final String dbPath;
   final String dbType;
+  final bool isSync;
 
-  DataBaseDownloadPage({this.isExistsDataBase, this.dbPath, this.dbType});
+  DataBaseDownloadPage(
+      {this.isExistsDataBase, this.dbPath, this.dbType, this.isSync});
 
   @override
   DataBaseDownloadPagepState createState() {
@@ -49,24 +51,6 @@ class DataBaseDownloadPage extends StatefulWidget {
 }
 
 class DataBaseDownloadPagepState extends State<DataBaseDownloadPage> {
-  final imgUrls = {
-    // Global
-    'global':
-        "https://github.com/violet-dev/db/releases/download/2020.07.07/hitomidata.7z",
-    // Korean
-    'ko':
-        "https://github.com/violet-dev/db/releases/download/2020.07.07/hitomidata-korean.7z",
-    // English
-    'en':
-        "https://github.com/violet-dev/db/releases/download/2020.07.07/hitomidata-english.7z",
-    // Japanese
-    'jp':
-        "https://github.com/violet-dev/db/releases/download/2020.07.07/hitomidata-japanese.7z",
-    // Chinese
-    'zh':
-        "https://github.com/violet-dev/db/releases/download/2020.07.07/hitomidata-chinese.7z",
-  };
-
   bool downloading = false;
   var baseString = "";
   var progressString = "";
@@ -83,8 +67,15 @@ class DataBaseDownloadPagepState extends State<DataBaseDownloadPage> {
   Future checkDownload() async {
     try {
       if ((await SharedPreferences.getInstance()).getInt('db_exists') == 1) {
-        await File((await SharedPreferences.getInstance()).getString('db_path'))
-            .delete();
+        if (await File(
+                (await SharedPreferences.getInstance()).getString('db_path'))
+            .exists())
+          await File(
+                  (await SharedPreferences.getInstance()).getString('db_path'))
+              .delete();
+        var dir = await getApplicationDocumentsDirectory();
+        if (await Directory('${dir.path}/data').exists())
+          await Directory('${dir.path}/data').delete(recursive: true);
       }
     } catch (e) {}
 
@@ -122,6 +113,8 @@ class DataBaseDownloadPagepState extends State<DataBaseDownloadPage> {
 
     try {
       var dir = await getApplicationDocumentsDirectory();
+      if (await File("${dir.path}/db.sql.7z").exists())
+        await File("${dir.path}/db.sql.7z").delete();
       Timer _timer = new Timer.periodic(
           Duration(seconds: 1),
           (Timer timer) => setState(() {
@@ -149,24 +142,39 @@ class DataBaseDownloadPagepState extends State<DataBaseDownloadPage> {
       _timer.cancel();
 
       setState(() {
-        downloading = false;
         baseString = Translations.instance.trans('dbdunzip');
+        print(baseString);
+        downloading = false;
       });
 
       var pp = new P7zip();
-      await pp.decompress(["${dir.path}/db.sql.7z"], path: "${dir.path}");
+      if (await Directory("${dir.path}/data2").exists())
+        await Directory("${dir.path}/data2").delete(recursive: true);
+      await pp.decompress(["${dir.path}/db.sql.7z"], path: "${dir.path}/data2");
+      if (await Directory('${dir.path}/data').exists())
+        await Directory('${dir.path}/data').delete(recursive: true);
+      await Directory("${dir.path}/data2").rename("${dir.path}/data");
+      if (await Directory("${dir.path}/data2").exists())
+        await Directory("${dir.path}/data2").delete(recursive: true);
 
       await File("${dir.path}/db.sql.7z").delete();
 
       await (await SharedPreferences.getInstance()).setInt('db_exists', 1);
-      await (await SharedPreferences.getInstance()).setString('db_path',
-          "${dir.path}/${imgUrls[widget.dbType].split('/').last.split('.')[0] + '.db'}");
+      await (await SharedPreferences.getInstance())
+          .setString('db_path', "${dir.path}/data/data.db");
       await (await SharedPreferences.getInstance())
           .setString('databasetype', widget.dbType);
       await (await SharedPreferences.getInstance()).setString('databasesync',
           UpdateSyncManager.rawlangDB[widget.dbType].item1.toString());
 
-      await indexing();
+      // await indexing();
+
+      if (widget.isSync != null && widget.isSync == true)
+        Navigator.pop(context);
+      else
+        setState(() {
+          baseString = Translations.instance.trans('dbdcomplete');
+        });
 
       return;
     } catch (e) {
